Don't delete new lyrics if there are no old ones.
svn path=/trunk/extragear/multimedia/amarok/; revision=475919

+++ src/collectiondb.cpp
@@ -1515,8 +1515,9 @@ CollectionDB::migrateFile( const QString &oldURL, const QString &newURL )
     query( QString( "DELETE FROM statistics WHERE url = '%1';" )
         .arg( escapeString( newURL ) ) );
 
-    query( QString( "DELETE FROM lyrics WHERE url = '%1';" )
-        .arg( escapeString( newURL ) ) );
+    if ( !getLyrics( oldURL ).isEmpty() )
+        query( QString( "DELETE FROM lyrics WHERE url = '%1';" )
+            .arg( escapeString( newURL ) ) );
     //  Migrate
     query( QString( "UPDATE tags SET url = '%1' WHERE url = '%2';" )
         .arg( escapeString( newURL ) )
